Random misc bugfixing and weird runtime error elimination

+++ mobs.h
@@ -36,9 +36,9 @@ typedef struct
   char resetonnext;
   int initialtimetonextframe;
   int timetonextframe;
-  int xmoverate;
+  double xmoverate;
   int xmoveamount;
-  int ymoverate;
+  double ymoverate;
   int ymoveamount;
 } mob;
 
@@ -56,6 +56,6 @@ void remove_mob(int);
 void mob_set_animation(mob* m, unsigned int animation, unsigned int startframe, unsigned int targetframe, unsigned int framesperframe, char looping);
 void  animate_mobs();
 void  move_mobs();
-void  mob_set_movement(mob* m, int xam, int xrate, int yam, int yrate);
+void  mob_set_movement(mob* m, int xam, double xrate, int yam, double yrate);
 void  mob_move_all (mob * m, int xtiles, int ytiles, int frames);
 #endif
